게시판 Toolbar 컴포넌트 화, 사이드바 컴포넌트 작업 완료
