+++ src/app/(route)/(community)/_components/RightSideBar.tsx
@@ -1,0 +1,9 @@
+export const RightSideBar = () => {
+  return (
+    <div className="w-[288px] minh-[808px]">
+      <div className="sticky top-[314px]  w-full min-h-[92px] bg-[#FFFFFF]">
+        뉴스
+      </div>
+    </div>
+  );
+};
